Add explicit pagination and action polling helpers

+++ crates/cloud-sdk-hetzner/src/cloud/catalog.rs
@@ -287,7 +287,7 @@ impl<'a> CatalogListRequest<'a> {
             write_pair(output, &mut len, &mut first, "type", kind.as_api_str())?;
         }
         if let Some(page) = self.page {
-            write_pair_u64(output, &mut len, &mut first, "page", u64::from(page.get()))?;
+            write_pair_u64(output, &mut len, &mut first, "page", page.get())?;
         }
         if let Some(per_page) = self.per_page {
             write_pair_u64(
