+++ crates/cloud-sdk-hetzner/src/cloud/images.rs
@@ -185,7 +185,7 @@ impl ImageListRequest {
             writer.u64_pair("bound_to", id.get())?;
         }
         if let Some(page) = self.page {
-            writer.u64_pair("page", u64::from(page.get()))?;
+            writer.u64_pair("page", page.get())?;
         }
         if let Some(per_page) = self.per_page {
             writer.u64_pair("per_page", u64::from(per_page.get()))?;
