+++ crates/cloud-sdk-hetzner/src/cloud/servers/placement_groups.rs
@@ -178,7 +178,7 @@ impl<'a> PlacementGroupListRequest<'a> {
             writer.pair("label_selector", selector.as_str())?;
         }
         if let Some(page) = self.page {
-            writer.u64_pair("page", u64::from(page.get()))?;
+            writer.u64_pair("page", page.get())?;
         }
         if let Some(per_page) = self.per_page {
             writer.u64_pair("per_page", u64::from(per_page.get()))?;
