+++ crates/cloud-sdk-hetzner/src/dns/zones/actions.rs
@@ -161,7 +161,7 @@ impl ZoneActionListRequest {
             writer.u64_pair("id", id.get())?;
         }
         if let Some(page) = self.page {
-            writer.u64_pair("page", u64::from(page.get()))?;
+            writer.u64_pair("page", page.get())?;
         }
         if let Some(per_page) = self.per_page {
             writer.u64_pair("per_page", u64::from(per_page.get()))?;
